[refactor] 조회수 증가 관련 유지보수를 위한 리팩터링
