+++ src/main/java/com/example/demo/domain/board/service/component/ViewIncreaseCommand.java
@@ -1,0 +1,9 @@
+package com.example.demo.domain.board.service.component;
+
+import com.example.demo.domain.board.Repository.BoardRepository;
+import com.example.demo.domain.board.Repository.ViewRepository;
+
+
+public interface ViewIncreaseCommand {
+    void increaseView(Long boardId);
+}
